requirements.txt
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.5.0
requests>=2.31.0
transformers>=4.35.0
fastmcp>=2.14.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
torch>=2.0.0  # Required for transformers
accelerate>=0.25.0  # For GPU support
# Azure (if using)
azure-storage-blob>=12.0.0
azure-identity>=1.15.0
pandas>=2.0.0 
numpy>=1.24.0  